+++ src/vanguards/control.py
@@ -39,6 +39,6 @@ def try_close_circuit(controller, circ_id):
   if _CLOSE_CIRCUITS:
     try:
       controller.close_circuit(circ_id)
-      plog("NOTICE", "We force-closed circuit "+str(circ_id))
+      plog("INFO", "We force-closed circuit "+str(circ_id))
     except stem.InvalidRequest as e:
       plog("INFO", "Failed to close circuit "+str(circ_id)+": "+str(e.message))