+++ CHANGES.rst
@@ -20,6 +20,7 @@ next
 - Demote Tor Bug #29699 dropped cell NOTICE to INFO. #66
 - Close circuits before logging about the issue. #67.
 - Always close PATH_BIAS_TESTING circutis with invalid cells. #68
+- Eliminate clearly invalid cases of allowed dropped cells. #69
 
 
 0.3.1